Update the nix flake to only include necessary packages
This flake contained several references to a dotfile manager that
the configuration appears to have been copied from. Similarly, it
looks like there are several packages that are not required by
jrsonnet that are included due to this copy/paste.

This commit
- Updates names/descriptions to "jrsonnet"
- Removes all dependencies except for rust
- Updates nix related utilities to the latest versions to support
  builds on 'aarch64-darwin'
- Updates rust channel to stable — the nightly release is not
  necessary for the features in use

Changes were confirmed to work with `nix flake check`, `nix build`,
and `nix develop`. `cargo test` and `cargo build` were tested in
the development environment. Tests were confined to aarch64-darwin
and aarch64-linux.